[WIP] Introduce paragraph count summary
Let the users use only the n-first paragraphs of the article as article
summary.

The advantage of this approach is that we avoid the random word count
ellipsis that will cut content in pieces, while not having to copy the
first paragraph of the article into the article's summary metadata.

Currently missing in this commit:

- Update documentation about negative value for paragraph count.