templates/tasks.py: livereload: automatically open browser
When the livereload target is invoked, a web brower will be
automatically opened pointing to the locally-served website.

In case of no web browser can be found by the module, the open call
returns False but no exception is raised. This means that it is still
possible to call livereload on a remote machine and access it without
any error being triggered.

Signed-off-by: Romain Porte <microjoe@microjoe.org>